test(sidebar): move git-lane pins beside their implementation
