+++ crates/rimz/src/sidebar/enrich/tests/mod.rs
@@ -7,13 +7,9 @@ use crate::pane::{RuntimeOwner, RuntimeOwnerKind};
 use crate::remote::link::{LinkStats, LinkStatsFile};
 use crate::sidebar::refresh::AccountsCache;
 use crate::sidebar::refresh::daemon_reap::write_codex_daemon_reap;
-use crate::sidebar::refresh::git_stats::{
-    DiffStatsCache, DiffStatsCacheEntry, WorktreeRootsCache, focused_worktree_paths,
-    hot_worktree_paths, needed_worktree_paths,
-};
+use crate::sidebar::refresh::git_stats::{DiffStatsCache, DiffStatsCacheEntry, WorktreeRootsCache};
 use crate::sidebar::refresh::{CodexDaemonReap, read_codex_daemon_reap};
 use crate::sidebar::test_support::{activity_row, pane, root_agent, worktree_group};
-use crate::sidebar::timing::GIT_ACTIVITY_WINDOW;
 use crate::store::snapshot::LinkTier;
 use crate::utils::time::unix_now_ms;
 use jiff::SignedDuration;
@@ -25,7 +21,6 @@ mod cohort;
 mod frame;
 mod git;
 mod labels;
-mod paths;
 mod spend;
 
 fn runtime() -> (tempfile::TempDir, RuntimePaths, SidebarSnapshot) {
